+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
     *Manuel Puyol*
 
+* Unify contributing guidelines.
+
+    *Kate Higa*
+
 * Rerun flaky system tests.
 
     *Manuel Puyol*
@@ -39,6 +43,10 @@
     *Manuel Puyol*
 
 ### Misc
+
+* Replace Classify::Spacing class with pre-generated mappings.
+
+  *Jon Rohan*
 
 * Add linter suggestions for `Button` component.