+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Restrict usage of padding system arguments on BorderBox, recommending use of `padding` density instead.
 
     *Joel Hawksley*
+
 ### Misc
 
 * Add autocorrect for button linters.