+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## main
 
+### Updates
+
+* Allow copying from elements using `for` in `ClipboardCopy`.
+
+    *Manuel Puyol*
+
+### Breaking changes
+
+* Remove `label` argument in favor of `aria-label` in `ClipboardCopy`.
+
+    *Manuel Puyol*
+
 ### Misc
 
 * Add autocorrect for button linters.